scripts/requirements.txt
markitdown[all]>=0.0.1
requests>=2.31.0
pyyaml>=6.0
pandas>=2.0.0
openpyxl>=3.1.0
